PORTAFOLIO/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
backports.tarfile==1.2.0
bidict==0.23.1
blinker==1.8.2
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
docutils==0.20.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.110.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.5.0
lazy_loader==0.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.21
numpy==1.24.4
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
peewee==3.18.1
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.3.6
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
readme_renderer==43.0
redis==5.3.0
reflex>=0.6.6.post1
reflex-chakra==0.5.10
reflex-hosting-cli==0.1.32
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.24
starlette==0.37.2
starlette-admin==0.14.1
tabulate==0.9.0
tomli==2.2.1
tomlkit==0.13.2
twine==5.1.1
typer==0.15.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.2
wsproto==1.2.0
zipp==3.20.2
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary==2.9.9
